Address comments
